Fix delete category test

+++ test/test_services/test_categories_service.py
@@ -66,33 +66,37 @@ def test_delete_cat_with_tr(db_env):
 
     with UnitOfWork(db_path) as uow:
         UtilTest.init_database(uow)
-        _, cat_list, tr_list = UtilTest.fill_user_cat_tr(uow, n_prim=15, n_tr=500)
+        _, cat_tot_list, tr_list = UtilTest.fill_user_cat_tr(uow, n_prim=15, n_tr=500)
 
     n = 20
     for _ in range(n):
-        cat = random.choice(cat_list)
+        cat = random.choice(cat_tot_list)
         cat_list = cat_service.get_secondary_list(uow, cat.id_user, None, None)
-        cat = random.choice(cat_list)
+        if cat_list:
+            cat = random.choice(cat_list)
 
-        begin_date = date(cat.year - 1, 1, 1)
-        end_date = date(cat.year + 1, 1, 1)
-        tr_list = tr_service.get_transaction(
-            uow=uow,
-            id_user=cat.id_user,
-            begin_date=begin_date,
-            end_date=end_date,
-            tr_type=cat.category_type,
-            primary=cat.primary,
-            secondary=cat.secondary,
-        )
+            begin_date = date(cat.year - 1, 1, 1)
+            end_date = date(cat.year + 1, 1, 1)
+            tr_list = tr_service.get_transaction(
+                uow=uow,
+                id_user=cat.id_user,
+                begin_date=begin_date,
+                end_date=end_date,
+                tr_type=cat.category_type,
+                primary=cat.primary,
+                secondary=cat.secondary,
+            )
 
-        tr_list = [tr for tr in tr_list if tr.tr_date.year == cat.year]
-        tr_get = cat_service.delete(uow, cat.id_secondary)
-        compare_tr(tr_list, tr_get)
+            tr_list = [tr for tr in tr_list if tr.tr_date.year == cat.year]
+            tr_get = cat_service.delete(uow, cat.id_secondary)
+            if tr_list == []:
+                assert tr_get is None
+            else:
+                compare_tr(tr_list, tr_get)
 
-        # cat are not deleted
-        cat_get = cat_service.get(uow, cat.id_user, cat.year, cat.category_type)
-        assert cat in cat_get
+                # cat is not deleted
+                cat_get = cat_service.get(uow, cat.id_user, cat.year, cat.category_type)
+                assert cat in cat_get
 
 
 # --- Utils
